requirements.txt
alabaster==0.7.12
astroid @ file:///C:/ci/astroid_1613501147567/work
async-generator==1.10
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
Babel @ file:///tmp/build/80754af9/babel_1607110387436/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247269562/work
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
comtypes==1.1.10
contextvars==2.4
cryptography @ file:///C:/ci/cryptography_1616769514516/work
decorator @ file:///tmp/build/80754af9/decorator_1617916966915/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dlib @ file:///D:/bld/dlib_1617441818780/work
docopt==0.6.2
docutils @ file:///C:/ci/docutils_1617624823126/work
entrypoints==0.3
face-recognition @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_1600102073710/work
face-recognition-models @ file:///home/conda/feedstock_root/build_artifacts/face_recognition_models_1595707436534/work
googletrans==3.1.0a0
gTTS==2.2.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imagesize @ file:///home/ktietz/src/ci/imagesize_1611921604382/work
immutables==0.15
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877476772/work
ipykernel==5.5.3
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort @ file:///tmp/build/80754af9/isort_1616355431277/work
jedi==0.18.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1612213139570/work
Js2Py==0.71
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///C:/ci/jupyter_core_1612213536848/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
lazy-object-proxy @ file:///C:/ci/lazy-object-proxy_1616529300868/work
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.0.24
mysql-connector-python-rf==2.2.2
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
numpy @ file:///D:/bld/numpy_1610324703282/work
numpydoc @ file:///tmp/build/80754af9/numpydoc_1605117425582/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.5.1.48
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandocfilters @ file:///C:/ci/pandocfilters_1605120535071/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
pipwin==0.5.1
playsound==1.2.2
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil @ file:///C:/ci/psutil_1612298125479/work
PyAudio @ file:///C:/Users/sintg/pipwin/PyAudio-0.2.11-cp36-cp36m-win_amd64.whl
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes @ file:///tmp/build/80754af9/pyflakes_1617200973297/work
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyjsparser==2.7.1
PyJWT==1.7.1
pylint @ file:///C:/ci/pylint_1617135975205/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pypiwin32==223
PyPrind==2.11.3
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
pySmartDL==1.3.4
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pyttsx3==2.90
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==300
pyzmq==22.0.3
QtAwesome @ file:///tmp/build/80754af9/qtawesome_1615991616277/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1616775094278/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rfc3986==1.4.0
rope @ file:///tmp/build/80754af9/rope_1602264064449/work
six @ file:///C:/ci/six_1605187303045/work
sniffio==1.2.0
snowballstemmer @ file:///tmp/build/80754af9/snowballstemmer_1611258885636/work
soupsieve==2.2.1
SpeechRecognition==3.8.1
Sphinx @ file:///tmp/build/80754af9/sphinx_1616268783226/work
sphinxcontrib-applehelp @ file:///home/ktietz/src/ci/sphinxcontrib-applehelp_1611920841464/work
sphinxcontrib-devhelp @ file:///home/ktietz/src/ci/sphinxcontrib-devhelp_1611920923094/work
sphinxcontrib-htmlhelp @ file:///home/ktietz/src/ci/sphinxcontrib-htmlhelp_1611920974801/work
sphinxcontrib-jsmath @ file:///home/ktietz/src/ci/sphinxcontrib-jsmath_1611920942228/work
sphinxcontrib-qthelp @ file:///home/ktietz/src/ci/sphinxcontrib-qthelp_1611921055322/work
sphinxcontrib-serializinghtml @ file:///home/ktietz/src/ci/sphinxcontrib-serializinghtml_1611920755253/work
spyder==3.2.8
spyder-kernels==1.10.0
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets==4.3.3
twilio==6.57.0
typed-ast @ file:///C:/ci/typed-ast_1610466539988/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
tzlocal==2.1
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
win10toast==0.9
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
